pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agroeye-decision-maker"
version = "0.1.0"
description = "Production-ready greenhouse controller for AGC 2019 dataset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "joblib>=1.3",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "pydantic>=2.6"
]

[project.optional-dependencies]
onnx = ["onnxruntime>=1.17", "skl2onnx>=1.16"]
boost = ["xgboost>=2.0", "lightgbm>=4.3"]
torch = ["torch>=2.2"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
